# expression: implement vectorized evaluation for builtinDayNameSig